% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{add_data}
\alias{add_data}
\title{Add Experimental Data}
\usage{
add_data(experiment, new_response)
}
\arguments{
\item{experiment}{GADGET experiment object}

\item{new_response}{matrix where each row is the response for the corresponding row of the current batch design computed by GADGET}
}
\description{
Merges newly collected data with the current stage of the GADGET experiment.
}

