% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ale.R
\name{ale_numeric_feature}
\alias{ale_numeric_feature}
\title{ALE for a single numeric feature.}
\usage{
ale_numeric_feature(
  model,
  data,
  X,
  feature,
  target_feature_name,
  n_intervals = 10,
  predict_fun = NULL
)
}
\arguments{
\item{model}{(`any`) \cr
Fitted model. See \code{\link{calculate_ale}}.}

\item{data}{(`data.frame()` or `data.table()`) \cr
Training data.}

\item{X}{(`data.frame()` or `data.table()`) \cr
Features (excl. target).}

\item{feature}{(`character(1)`) \cr
Feature name.}

\item{target_feature_name}{(`character(1)`) \cr
Target name.}

\item{n_intervals}{(`integer(1)`) \cr
Number of intervals.}

\item{predict_fun}{(`function()` or `NULL`) \cr
Prediction function.}
}
\value{
(`data.table()`) \cr
  ALE data with \code{row_id}, \code{feat_val}, \code{d_l}, \code{interval_index}, etc.
}
\description{
ALE for a single numeric feature.
}
\keyword{internal}
