% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bastos-ohagan-09.R
\name{bo09_toy}
\alias{bo09_toy}
\title{Two-Input Toy Model from Bastos and O'Hagan 2009}
\usage{
bo09_toy(x)
}
\arguments{
\item{x}{vector from [0,1] x [0,1]}
}
\value{
scalar
}
\description{
A toy model for the demonstration of Gaussian process validation plot and statistics.
}
\examples{
bo09_toy(c(0.5,0.25))

}
\references{
Bastos, L. S., & O’hagan, A. (2009). Diagnostics for gaussian process emulators. Technometrics, 51(4), 425–438. https://doi.org/10.1198/TECH.2009.08019
}
