% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{build_node_title}
\alias{build_node_title}
\title{Build node title for ALE or PD tree plots.}
\usage{
build_node_title(node, depth_idx, tree, style = c("ale", "pd"))
}
\arguments{
\item{node}{(`Node`) \cr
Node object.}

\item{depth_idx}{(`integer(1)`) \cr
Depth index.}

\item{tree}{(`list()`) \cr
Depth-based list of Node objects.}

\item{style}{(`character(1)`) \cr
\code{"ale"} or \code{"pd"} for format variant.}
}
\value{
(`character(1)`) \cr
  Title string.
}
\description{
Build node title for ALE or PD tree plots.
}
\keyword{internal}
