% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ale_heterogeneity.R
\name{calculate_ale_heterogeneity_cpp}
\alias{calculate_ale_heterogeneity_cpp}
\title{Calculate ALE Heterogeneity}
\usage{
calculate_ale_heterogeneity_cpp(Y)
}
\arguments{
\item{Y}{(`list()` or `data.frame()`) \cr
ALE effect data.}
}
\value{
(`numeric()`) \cr
  Heterogeneity value(s): vector per feature when Y is list, single value when Y is data.frame.
}
\description{
Calculate ALE Heterogeneity
}
\keyword{internal}
