% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_tree_splitting.R
\name{compute_data_for_ice_splitting}
\alias{compute_data_for_ice_splitting}
\title{Prepare centered ICE matrices and covariate data for split search}
\usage{
compute_data_for_ice_splitting(effect, testdata, Z)
}
\arguments{
\item{effect}{An object returned by \pkg{iml}\::\code{FeatureEffect}
(with \code{method = "ice"}). Must contain slots
\code{\$results} and \code{\$features}.}

\item{testdata}{The original data set passed to
\code{FeatureEffect$new()} (or a subset of it) as a
\code{data.frame}/\code{data.table}.}

\item{Z}{Character vector of column names to be used as splitting features}
}
\value{
A named \code{list} with three elements:
  \describe{
    \item{\code{X}}{A \code{data.table} containing the selected
      covariates, coercing factors to numeric when needed.}
    \item{\code{Y}}{A list of \code{data.table}s, one per
      feature, each in wide ICE matrix form and centered by row
      means. Each row corresponds to a ice curve}
    \item{\code{grid}}{A list; each element is a character vector of
      column names (i.e.\ grid points) used for the corresponding
      ICE matrix in \code{Y}.}
  }
}
\description{
Prepare centered ICE matrices and covariate data for split search
}
\keyword{internal}
