% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mi_experiment.R
\name{create_mi_experiment}
\alias{create_mi_experiment}
\title{Creates a Mutual Information Experiment object}
\usage{
create_mi_experiment(MI = list("knn", 6), design, response, posterior_sampler,
  posterior_parms = NULL, predictive_sampler, predictive_parms = NULL,
  lower_bound, upper_bound, batch = 1, explore_budget = c(10, 10),
  design_budget = 10, gp_options = list(), simulation = NULL,
  simulation_parms = NULL)
}
\arguments{
\item{MI}{mutual information algorithm to be used as a design criteria and additional parameters}

\item{design}{list of design points that data have been collected}

\item{response}{list of responses that have been collected}

\item{posterior_sampler}{function that returns a sample from the posterior distribution}

\item{posterior_parms}{list of parameters to pass to posterior_sampler}

\item{predictive_sampler}{function that produces a sample from the predictive distribution}

\item{predictive_parms}{list of parameters to pass to predictive_sampler}

\item{lower_bound}{vector containing the lower bound(s) of the design space}

\item{upper_bound}{vector containing the upper bound(s) of the design space}

\item{batch}{number of design points to be found per iteration}

\item{explore_budget}{vector contain the number of initial space filling points and the number of EQI exploration points}

\item{design_budget}{number of batches that GADGET will compute}

\item{gp_options}{options that specify the type of Gaussian process to fit to the mutual information surface}

\item{simulation}{computer simulation being explored}

\item{simulation_parms}{extra parameters to call the computer simulation with}
}
\description{
A GADGET experiment object is created with can be processed to sequentially to
pick new batches of design points. All needed functions are collected together
so that the user does not need to handle any function calls.
}

