% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_pd.R
\name{create_node_title}
\alias{create_node_title}
\title{Create node title for PD plot (internal).}
\usage{
create_node_title(node, depth_idx, tree)
}
\description{
Create node title for PD plot (internal).
}
\keyword{internal}
