% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ut_experiment.R
\name{create_ut_experiment}
\alias{create_ut_experiment}
\title{Creates a utility based Experiment object}
\usage{
create_ut_experiment(utility, design, response, posterior_sampler,
  posterior_parms = NULL, lower_bound, upper_bound, batch = 1,
  explore_budget = c(10, 10), design_budget = 10, gp_options = list(),
  simulation = NULL, simulation_parms = NULL)
}
\arguments{
\item{utility}{function which evaluates the utility of a proposed design}

\item{design}{list of design points that data have been collected}

\item{response}{list of responses that have been collected}

\item{posterior_sampler}{function that returns a sample from the posterior distribution}

\item{posterior_parms}{list of parameters to pass to posterior_sampler}

\item{lower_bound}{vector containing the lower bound(s) of the design space}

\item{upper_bound}{vector containing the upper bound(s) of the design space}

\item{batch}{number of design points to be found per iteration}

\item{explore_budget}{vector contain the number of initial space filling points and the number of EQI exploration points}

\item{design_budget}{number of observation the experimental design should contain}

\item{gp_options}{options that specify the type of Gaussian process to fit to the expected utility surface}

\item{simulation}{computer simulation being explored}

\item{simulation_parms}{extra parameters to call the computer simulation with}
}
\description{
Creates a utility based Experiment object
}

