% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_experiment.R
\name{design_experiment}
\alias{design_experiment}
\title{Design Optimal Experiment using Gaussian Process Optimization}
\usage{
design_experiment(design_criterion, stochastic = TRUE, lower_bound,
  upper_bound, init_budget = 10, optim_budget = 10,
  gp_options = list(formula = ~1, kernel = "matern5_2", optimizer =
  "gen"), genoud_options = list(pop.size = 1000, max.generations = 100,
  wait.generations = 10), diagnostics = 1, verbose = 1,
  max_augment = 10, cluster = NULL)
}
\arguments{
\item{design_criterion}{(see details)}

\item{stochastic}{}

\item{lower_bound}{}

\item{upper_bound}{}

\item{init_budget}{}

\item{optim_budget}{}

\item{gp_options}{}

\item{genoud_options}{}

\item{diagnostics}{}

\item{verbose}{}

\item{max_augment}{}

\item{cluster}{}
}
\description{
If it is a deterministic criterion then the minimum is taken over the entire set of dc runs instead of taking the last one. this is the simpliest implimentation of the method by Weaver et. al. this function will essentially be the same as run_stage this will perform the essential functions of Weaver's paper actually this can be used to optimize any black box stochastic function just put the objective function as the design_criterion and the resulting design will be the argmin of the
}
\details{

}
\examples{

#static d-optimal
#see vignettes


}
