% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{endGadget}
\alias{endGadget}
\title{Stop the active gadget simulation run}
\usage{
endGadget()
}
\description{
This function stops a currently active gagdet simulation and release all the allocated memory.
}
