% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_split_data_utils.R
\name{ensure_factors}
\alias{ensure_factors}
\title{Ensure character columns are factors.}
\usage{
ensure_factors(data, cols)
}
\arguments{
\item{data}{(`data.frame()` or `data.table()`) \cr
Data to modify.}

\item{cols}{(`character()`) \cr
Column names to ensure as factor.}
}
\value{
(`data.frame()` or `data.table()`) \cr
  Modified data (by reference for data.table).
}
\description{
Ensure character columns are factors.
}
\keyword{internal}
