% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_node_by_id.R
\name{find_node_by_id}
\alias{find_node_by_id}
\title{Find node by id in a flat list.}
\usage{
find_node_by_id(node_list, id)
}
\arguments{
\item{node_list}{(`list()`) \cr
List of Node objects (each with an \code{id} field).}

\item{id}{(`integer(1)`) \cr
Node id to look up.}
}
\value{
(`Node` or `NULL`) \cr
  The node with matching \code{id}, or \code{NULL} if not found.
}
\description{
Given node_list and id: loops and returns first node with \code{node$id == id}, or NULL.
}
\keyword{internal}
