% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_parent_by_id.R
\name{find_parent_by_id}
\alias{find_parent_by_id}
\title{Find node by id in a flat list. Given node.list and id: loops and returns first node with \code{node$id == id}, or NULL.}
\usage{
find_parent_by_id(node.list, id)
}
\arguments{
\item{node.list}{List of node objects (each with an \code{id} field).}

\item{id}{Scalar; node id to look up.}
}
\value{
The node with \code{id} or \code{NULL}.
}
\description{
Find node by id in a flat list. Given node.list and id: loops and returns first node with \code{node$id == id}, or NULL.
}
\keyword{internal}
