% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget-internal.R
\name{gadget-internal}
\alias{gadget-internal}
\alias{ale_sweep_cpp}
\alias{calculate_ale_heterogeneity_list_cpp}
\alias{calculate_ale_heterogeneity_single_cpp}
\alias{re_mean_center_ice_cpp}
\alias{search_best_split_cpp}
\alias{search_best_split_point_cpp}
\alias{dL}
\alias{interval.index}
\alias{level}
\alias{x}
\alias{x.grid}
\alias{x.left}
\alias{x.right}
\alias{y}
\title{Internal C++ helpers and package symbols}
\description{
Functions and symbols used internally by the package. Not intended for direct use.
}
\keyword{internal}
