% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GADGET.R
\docType{package}
\name{GADGET}
\alias{GADGET}
\alias{GADGET-package}
\title{Gaussian Process Approximations for Designing Experiment}
\description{
The \code{GADGET} package offers
}
\section{References}{

Gao, S., Ver Steeg G., & Galstyan A. (2015). Efficient estimation of mutual information for strongly dependent variables. Artificial Intelligence and Statistics: 277-286.
}

\author{
Isaac Michaud
}
