% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget_internal.R
\name{gadget_internal}
\alias{gadget_internal}
\alias{ale_sweep_cpp}
\alias{calculate_ale_heterogeneity_list_cpp}
\alias{calculate_ale_heterogeneity_single_cpp}
\alias{re_mean_center_ice_cpp}
\alias{search_best_split_cpp}
\alias{default_predict_fun}
\alias{predict_newdata_fast_dispatch}
\alias{extract_numeric_prediction}
\alias{has_predict_method}
\alias{cpp_pd_stack_newdata}
\alias{risk_from_stats}
\alias{assert_ale_effect_list}
\alias{d_l}
\alias{interval_index}
\alias{level}
\alias{x}
\alias{x_grid}
\alias{x_left}
\alias{x_right}
\alias{y}
\title{Internal C++ helpers and package symbols}
\description{
Functions and symbols used internally by the package. Not intended for direct use.
}
\keyword{internal}
