% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getEcosystemInfo}
\alias{getEcosystemInfo}
\title{Get the simulation ecosystem objects}
\usage{
getEcosystemInfo()
}
\description{
This function returns a vector of information about the running gadget
simulation's stocks, fleet, state, and time information.
}
