% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getFleetNo}
\alias{getFleetNo}
\title{Get the internal fleet identifier}
\usage{
getFleetNo(fleetName)
}
\arguments{
\item{fleetName}{String name of the fleet}
}
\value{
an internal fleet number
}
\description{
Converts a fleet name into the interanl fleet identifier in a running simulation
}
