% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getLenGrpIdx}
\alias{getLenGrpIdx}
\title{Get a stock's length group index}
\usage{
getLenGrpIdx(stockName, len)
}
\arguments{
\item{stockName}{a string of the stock name}

\item{len}{the source length}
}
\value{
a length group index identifier
}
\description{
This function converts a length measurement into a stock's internal length group index
}
