% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getPredation}
\alias{getPredation}
\title{Get the predatory (eating/catch) statistics of a stock/fleet subject}
\usage{
getPredation(x, type)
}
\arguments{
\item{x}{Stock or Fleet name}

\item{type}{Must be either "stock" or "fleet"}
}
\value{
A matrix of predator-prey statistics of an subject inside the list(catch = ...) object
}
\description{
This function produces similar output as the standalone Gadget's printPredatorPrey type output
}
\examples{
\dontrun{
if(interactive()){
 exPath <- loadExample()
 initGadget(exPath, "refinputfile")
 stat <- getPredation("comm", "fleet")
 print(stat$catch$had)
 endGadget()
 }
}
}
