% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getStockNo}
\alias{getStockNo}
\title{Get the internal stock identifier}
\usage{
getStockNo(stockName)
}
\arguments{
\item{stockName}{String name of the stock}
}
\value{
an internal stock identifier
}
\description{
Converts a stock name into an internal stock identifier in a running simulation
}
