% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_plot.R
\name{gp_plot}
\alias{gp_plot}
\title{Plot 1-D DiceKriging Gaussian Process Model}
\usage{
gp_plot(model, inputs, plot_data = FALSE)
}
\arguments{
\item{model}{a km GP model from DiceKriging}

\item{inputs}{a vector of values at which to evaluate the GP model}

\item{plot_data}{logical, should the data used to generate the GP be plotted too?}
}
\description{
Plots simple Gaussian process (GP) models with 95% credible bands.
}
