% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_plot_data_ale.R
\name{mean_center_ale}
\alias{mean_center_ale}
\title{Internal ALE curve computation}
\usage{
mean_center_ale(feat, mean.center = TRUE)
}
\arguments{
\item{feat}{data.table with per-interval ALE derivatives and metadata.}

\item{mean.center}{Logical. Whether to mean-center the resulting ALE
curve (default \code{TRUE}).}
}
\description{
Given ALE data.table for one feature and mean.center: cumsums dL by interval; 
optionally subtracts global mean. Returns data.table with x.grid and .value (cumulative ALE).
}
