% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plots.R
\name{plot_tree_structure_with_ggraph}
\alias{plot_tree_structure_with_ggraph}
\title{Plot Tree Structure with ggraph}
\usage{
plot_tree_structure_with_ggraph(tree)
}
\arguments{
\item{tree}{A list of lists representing a tree structure.}
}
\value{
A \code{ggplot} object (of class \code{ggraph}) representing the tree layout.
  Typically used for visualization purposes.
}
\description{
Plot Tree Structure with ggraph
}
