% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_diagnostics.R
\name{predictive_errors}
\alias{predictive_errors}
\title{Title}
\usage{
predictive_errors(design, response, model, plot = TRUE, type = "SK")
}
\arguments{
\item{design}{}

\item{response}{}

\item{model}{}

\item{plot}{}

\item{type}{}
}
\description{
Title
}
\examples{
# simple iid normal case
x        <- matrix(runif(20,-1.5,1.5),ncol=1) 
y        <- matrix(rnorm(20),ncol = 1)
my_model <- km(formula=~1,design=x,response=y,covtype='matern5_2',optim.method='BFGS',nugget.estim=FALSE)
# validation data
x        <- matrix(runif(20,-1,1),ncol=1)
y        <- matrix(rnorm(20),ncol = 1)
plotGP(my_model,c(-2,2),0.1)
predictive_errors(matrix(x,ncol=1),matrix(y,ncol=1),my_model)
f <- function(x) {
 x^2 
}
#fit model 
x        <- matrix(runif(20,-1.5,1.5),ncol=1) 
y        <- matrix(f(x),ncol = 1)
my_model <- km(formula=~.^2,design=x,response=y,covtype='matern5_2',optim.method='BFGS',nugget.estim=FALSE)
#validate model
x        <- matrix(runif(20,-1.5,1.5),ncol=1) 
y        <- matrix(f(x),ncol = 1)
plotGP(my_model,c(-5,5),0.1)
predictive_errors(matrix(x,ncol=1),matrix(y,ncol=1),my_model, type = "SK")
predictive_errors(matrix(x,ncol=1),matrix(y,ncol=1),my_model, type = "UK")
#y = matrix(f(x),ncol = 1)
}
