% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{preprocess_node_data}
\alias{preprocess_node_data}
\title{Preprocess PD node data by depth.}
\usage{
preprocess_node_data(tree, Y, grid_total, mean_center)
}
\arguments{
\item{tree}{(`list()`) \cr
Depth-based list of Node objects.}

\item{Y, grid_total}{(`list()`) \cr
Effect matrices and grid.}

\item{mean_center}{(`logical(1)`) \cr
Whether to mean-center.}
}
\value{
(`list()`) \cr
  Processed data per depth.
}
\description{
Preprocess PD node data by depth.
}
\keyword{internal}
