% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GADGET.R
\name{print_logo}
\alias{print_logo}
\title{Print ASCII GADGET Logo}
\usage{
print_logo()
}
\value{
Nothing
}
\description{
Prints the GADGET ASCII logo in the center of the console.
}
\examples{
print_logo()
}
\keyword{internal}
