% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GADGET.R
\name{print_separate}
\alias{print_separate}
\title{Print Separator}
\usage{
print_separate()
}
\value{
Nothing
}
\description{
Prints a string of dashes to separate output.
}
\examples{
print_separate()
}
\keyword{internal}
