% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GADGET.R
\name{print_seperator}
\alias{print_seperator}
\title{Print Seperator}
\usage{
print_seperator()
}
\value{
Nothing
}
\description{
Prints a string of dashes to seperate output.
}
\examples{
print_seperator()
}
\keyword{internal}
