% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mi_experiment.R
\name{run_mi_experiment}
\alias{run_mi_experiment}
\title{Run GADGET mutual information experiment}
\usage{
run_mi_experiment(experiment, verbose = FALSE)
}
\arguments{
\item{experiment}{GADGET mutual information experiment object}

\item{verbose}{logical. If FALSE extra output is suppressed.}
}
\description{
Sequentially designs and executes an experiment using the mutual information
design criterion.
}

