% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run simulation}
\usage{
run_simulation(experiment)
}
\arguments{
\item{experiment}{GADGET experiment object}
}
\description{
A computer experiment will involve an associated compute code that needs to be
run to collect new data. This function automatically collects new data to allow
GADGET to continue sequentially designing and running an experiment.
}

