% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stage.R
\name{run_stage}
\alias{run_stage}
\title{Run experiment stage}
\usage{
run_stage(experiment, design_criteria)
}
\arguments{
\item{experiment}{GADGET experiment object}

\item{design_criteria}{valid GADGET design criterion}
}
\description{
This function takes a GADGET experiment, computes the next batch of design
points. It uses a space-filling design to evalute the design criterion on a
parsimonious set of points. A Gaussian process is fit to these points and EQI
is used to sequentially find new points to update the Gaussian process to find
its minimum. The best batch design is then returned.
}

