% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ut_experiment.R
\name{run_ut_experiment}
\alias{run_ut_experiment}
\title{Run GADGET utility experiment}
\usage{
run_ut_experiment(experiment, verbose = FALSE)
}
\arguments{
\item{experiment}{GADGET utility experiment object}

\item{verbose}{logical. If FALSE extra output is suppressed.}
}
\description{
Sequentially designs and executes an experiment using the expected utility
criterion supplied by the user.
}

