% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget2flr.R
\name{runUntil}
\alias{runUntil}
\title{Instruct gadget to run until a specific year and produces FLR stock and index objects}
\usage{
runUntil(until, globalparams)
}
\arguments{
\item{until}{end year of the simulation}

\item{globalparams}{a list object of parameters to aid gadget to FLR objects generation (See this page
\code{vignette("gadget-flr", package = "gadgetr")} for an example)}
}
\description{
This function runs gadget until a specified year and automatically generate two 
FLR objects (FLStock and FLIndex) for each available stocks from the gadget run.
}
