% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{runYear}
\alias{runYear}
\title{Forward simulation for a year and collect the stats for all the stocks and fleets for that year}
\usage{
runYear()
}
\value{
list of fleets and stocks statistics (in matrix format) for the year run
}
\description{
Forward simulation for a year and collect the stats. For the stocks, the output consists of
stock number (stk), spawning stock biomass (ssb), recruitment number (rec), and predatory information
which are categorized by preys (stock names). Please note that both the stock number before and after the steps are
available. For the fleets, the output consists of only the predatory information which are categorized by
preys (stock names)
}
\examples{
\dontrun{
if(interactive()){
 exPath <- loadExample()
 initGadget(exPath, "refinputfile")
 stat <- runYear()
 print(stat)
 endGadget()
 }
}
}
