% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential_experiment.R
\name{sequential_experiment}
\alias{sequential_experiment}
\title{Design and Run Sequential Computer Experiment}
\usage{
sequential_experiment(design_criterion, stochastic, posterior_sampler,
  lower_bound, upper_bound, simulator, init_design = NULL,
  init_response = NULL, batch = 1, init_budget = 10, optim_budget = 10,
  design_budget = 10, gp_options = list(formula = ~1, kernel = "matern5_2",
  optimizer = "gen"), genoud_options = list(pop.size = 1000, max.generations =
  100, wait.generations = 10), diagnostics = 1, verbose = 1,
  max_augment = 10, cluster = NULL)
}
\arguments{
\item{posterior_sampler}{function that returns a sample from the posterior distribution}

\item{lower_bound}{vector containing the lower bound(s) of the design space}

\item{upper_bound}{vector containing the upper bound(s) of the design space}

\item{simulator}{computer simulator being explored (make it so that it is just a function of the design point)}

\item{batch}{number of design points per experiment stage (batch = 1 is sequential)}

\item{design_budget}{number of observation the experimental design should contain}

\item{gp_options}{options that specify the type of Gaussian process to fit to the expected utility surface}

\item{design_criteria}{function which evaluates the utility of a proposed design}

\item{design}{list of design points that data have been collected}

\item{response}{list of responses that have been collected}

\item{posterior_parms}{list of parameters to pass to posterior_sampler}

\item{explore_budget}{vector contain the number of initial space filling points and the number of EQI exploration points}

\item{simulator_parms}{extra parameters to call the computer simulator with}
}
\description{
Design and Run Sequential Computer Experiment
}
