% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_fill.R
\name{space_fill}
\alias{space_fill}
\title{Create Space-Filling Design}
\usage{
space_fill(experiment)
}
\arguments{
\item{experiment}{GADGET experiment object}
}
\value{
LHS design as matrix of size n by k*b
}
\description{
Each stage of a GADGET experiment starts by estimated the current design criterion
surface using a maximin latin hypercude sample (LHS). This function computes the multidimensional
LHS given the initial exploration budget, boundary constraints, and batch size.
}
\examples{
experiment <- list(lower=0,upper=50,batch=4,explore_budget=c(20,2),num_parms=1)
my_lhs     <- space_file(experiment)
}
