% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{stopEating}
\alias{stopEating}
\title{Instruct a stock to stop all of its predatory activities}
\usage{
stopEating(stockName)
}
\arguments{
\item{stockName}{a string of the stock name}
}
\value{
a vector of numeric error codes (0 is success, otherwise is error)
}
\description{
This function instructs a stock to stop all of its predatory activities.
This is equivalent with setting a stock's "doeseat" parameter to 0
}
