% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summay.GADGET_MI_EXP.R
\name{summary.GADGET_MI_EXP}
\alias{summary.GADGET_MI_EXP}
\title{Summary of mutual information GADGET experiment}
\usage{
\method{summary}{GADGET_MI_EXP}(experiment, stage = NULL)
}
\arguments{
\item{experiment}{GADGET mutual information experiment object}

\item{stage}{the stage of the GADGET experiment to display}
}
\description{
Prints a summary of a mutual information based GADGET experiment.
}

