% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.GADGET_UT_EXP.R
\name{summary.GADGET_UT_EXP}
\alias{summary.GADGET_UT_EXP}
\title{Summary of utility GADGET experiment}
\usage{
\method{summary}{GADGET_UT_EXP}(experiment, stage = NULL)
}
\arguments{
\item{experiment}{GADGET utility experiment object}

\item{stage}{the stage of the GADGET experiment to display}
}
\description{
Prints a summary of a utility based GADGET experiment.
}
