% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_split_data_utils.R
\name{take_cols}
\alias{take_cols}
\title{Subset columns from data.frame or data.table.}
\usage{
take_cols(data, cols)
}
\arguments{
\item{data}{(`data.frame()` or `data.table()`) \cr
Data.}

\item{cols}{(`character()`) \cr
Column names.}
}
\value{
(`data.frame()` or `data.table()`) \cr
  Subset of columns.
}
\description{
Subset columns from data.frame or data.table.
}
\keyword{internal}
