% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{updateAmount}
\alias{updateAmount}
\title{Update a total fleet catch amount}
\usage{
updateAmount(fleetName, year, step, area, value)
}
\arguments{
\item{fleetName}{a string of the fleet name}

\item{year}{a target year}

\item{step}{a target step}

\item{area}{a target area}

\item{value}{the target amount}
}
\value{
a vector of numeric error codes (0 is success, otherwise is error)
}
\description{
This function updates the catch amount of a totalfleet object.
}
\examples{
\dontrun{
if(interactive()){
 exPath <- loadExample()
 initGadget(exPath, "refinputfile")
 updateAmount("future", 1978, 1, 1, 123456)
 updateAmount("future", 1978, 3, 1, 567890)
 stat <- runStep()
 print(sum(stat$fleets$future$catch$had[, "biomassConsumed"]))
 print(aggregate(biomassConsumed ~ year + step, data = stat$fleets$future$catch$had, FUN=sum))
 endGadget()
 }
}
}
