% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget2flr.R
\name{updateFLStock}
\alias{updateFLStock}
\title{Update FLStock and FLIndex objects using a gadget simulation run output}
\usage{
updateFLStock(stockTitle, out, gadgetYear, fl_stock, fl_index, globalparams)
}
\arguments{
\item{stockTitle}{a string of gadget stock name}

\item{out}{gadget simulation output object. Normally is the output of [runYear()]}

\item{gadgetYear}{gadget year target}

\item{fl_stock}{an existing FLStock object. Normally produced by [runUntil()]}

\item{fl_index}{an existing FLIndex object. Normally produced by [runUntil()]}

\item{globalparams}{a list object of parameters to aid gadget to FLR objects generation (See this page
\code{vignette("gadget-flr", package = "gadgetr")} for an example)}
}
\description{
This function translates a gadget output object for a given \code{gadgetYear} 
year and \code{stockTitle} stock and inserts it to the FLR objects (\code{fl_stock} and \code{fl_index}).
}
