% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{updateRenewal}
\alias{updateRenewal}
\title{Update a stock's renewal parameters}
\usage{
updateRenewal(
  stockName,
  year,
  step,
  area,
  age,
  number,
  mean = NA,
  sdev = NA,
  alpha = NA,
  beta = NA,
  length = NA,
  meanWeight = NA
)
}
\arguments{
\item{stockName}{a string of the stock name}

\item{year}{the target year of renewal}

\item{step}{the target step of renewal}

\item{area}{the target area of renewal}

\item{age}{the age target of the stock}

\item{number}{the number parameter value}

\item{mean}{the mean parameter value (Normal Parametric Distribution only), Default: NA}

\item{sdev}{the sdev parameter value (Normal Parametric Distribution only), Default: NA}

\item{alpha}{the alpha parameter value (Normal Parametric Distribution only), Default: NA}

\item{beta}{the beta parameter value (Normal Parametric Distribution only), Default: NA}

\item{length}{the mean parameter value (Numerical Distribution only), Default: NA}

\item{meanWeight}{the weight parameter value (Numerical Distribution only), Default: NA}
}
\value{
a vector of numeric error codes (0 is success, otherwise is error)
}
\description{
This function updates a stock renewal ("recruitment") parameters, but only if
the stock is set to renew (i.e., doesrenew is set to 1 in the model). The function accepts a similar
input format as the original gadget renewal file. Note that only 2 modes of renewal (Normal 
Parametric and Numerical Distribution) that are supported for now.
}
\examples{
\dontrun{
if(interactive()){
 exPath <- loadExample()
 initGadget(exPath, "refinputfile")
 updateRenewal("had", 1978, 1, 1, 1, 123456, mean=16.41, sdev=2.25, alpha=8.85e-6, beta=3.0257)
 stat <- runStep(stockAfterStep = TRUE)
 print(sum(stat$stocks$had$rec$renew[, "renewalNumber"]))
 endGadget()
 }
}
}
