% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{updateSpawningRec}
\alias{updateSpawningRec}
\title{Update a stock's spawning parameters}
\usage{
updateSpawningRec(stockName, ...)
}
\arguments{
\item{stockName}{a string of the stock name}

\item{...}{the collection of parameters (e.g., if the recruitment function is "fecundity",
a vector of 5 parameters are required)}
}
\value{
a vector of numeric error codes (0 is success, otherwise is error)
}
\description{
This function updates a stock spawning recruitment parameters, but only if
the stock is set to spawn (i.e., doesspawn is set to 1 in the model). All spawning recruitment
functions are supported. However, it is the user responsibility to give the correct
parameters that correspond to the stock's spawning recruitment functions. See the
gadget user guide for the detailed information about the different functions and their parameters
}
