% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ale.R
\name{ale_numeric_feature}
\alias{ale_numeric_feature}
\title{ALE for a single numeric feature (internal)}
\usage{
ale_numeric_feature(
  model,
  data,
  X,
  feature,
  target.feature.name,
  n.intervals = 10,
  predict.fun = NULL
)
}
\arguments{
\item{model, data, X, feature, target.feature.name, n.intervals, predict.fun}{See \code{\link{calculate_ale}}.}
}
\description{
ALE for a single numeric feature (internal)
}
\keyword{internal}
