% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ale_heterogeneity.R
\name{calculate_ale_heterogeneity_cpp}
\alias{calculate_ale_heterogeneity_cpp}
\title{Calculate ALE Heterogeneity}
\usage{
calculate_ale_heterogeneity_cpp(Y)
}
\arguments{
\item{Y}{List or data.frame. ALE effect data.}
}
\value{
Numeric. If Y is a list, returns a list of heterogeneity values for each feature.
  If Y is a data.frame, returns a single heterogeneity value.
}
\description{
Calculate ALE Heterogeneity
}
