% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{calculate_y_range}
\alias{calculate_y_range}
\title{Compute y-axis range for PD plots.}
\usage{
calculate_y_range(prepared_data, data, target_feature_name)
}
\arguments{
\item{prepared_data}{(`list()`) \cr
Prepared effect data.}

\item{data}{(`data.frame()`) \cr
Original data.}

\item{target_feature_name}{(`character(1)`) \cr
Target column.}
}
\value{
(`list()`) \cr
  \code{ymin}, \code{ymax}.
}
\description{
Compute y-axis range for PD plots.
}
\keyword{internal}
