% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_pd.R
\name{calculate_y_range}
\alias{calculate_y_range}
\title{Compute y-axis range for PD plots (internal).}
\usage{
calculate_y_range(prepared.data, data, target.feature.name, show.point)
}
\description{
Compute y-axis range for PD plots (internal).
}
\keyword{internal}
