% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{calculate_y_range_ale}
\alias{calculate_y_range_ale}
\title{Compute y-axis range for ALE plots.}
\usage{
calculate_y_range_ale(curves, data, target_feature_name)
}
\arguments{
\item{curves}{(`list()`) \cr
Output of \code{prepare_plot_data_ale}.}

\item{data}{(`data.frame()` or `NULL`) \cr
Original data (optional).}

\item{target_feature_name}{(`character(1)`) \cr
Target column.}
}
\value{
(`list()`) \cr
  \code{ymin}, \code{ymax}.
}
\description{
Compute y-axis range for ALE plots.
}
\keyword{internal}
