% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_ale.R
\name{calculate_y_range_ale}
\alias{calculate_y_range_ale}
\title{Compute y-axis range for ALE plots (effect + optional target data, with padding)}
\usage{
calculate_y_range_ale(curves, data, target.feature.name, show.point)
}
\description{
Compute y-axis range for ALE plots (effect + optional target data, with padding)
}
\keyword{internal}
