% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_pd.R
\name{create_plots_for_depth}
\alias{create_plots_for_depth}
\title{Create PD plots for one depth (internal).}
\usage{
create_plots_for_depth(
  tree,
  prepared.data,
  data,
  target.feature.name,
  depth_idx,
  nodes_to_render,
  color.ice,
  color.pd,
  show.plot,
  show.point,
  mean.center
)
}
\description{
Create PD plots for one depth (internal).
}
\keyword{internal}
