% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_plot_data_ale.R
\name{cumsum_na_as_zero}
\alias{cumsum_na_as_zero}
\title{Cumulative Sum with NA Handling}
\usage{
cumsum_na_as_zero(values)
}
\arguments{
\item{values}{Numeric vector. Values to be cumulatively summed;
any \code{NA} values are treated as \code{0}.}
}
\description{
Cumulative Sum with NA Handling
}
