% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aleStrategy.R
\name{default_predict_fun}
\alias{default_predict_fun}
\title{Default prediction function for mlr3-style models (internal).}
\usage{
default_predict_fun(model, data)
}
\arguments{
\item{model}{Fitted model.}

\item{data}{New data.}
}
\description{
Default prediction function for mlr3-style models (internal).
}
\keyword{internal}
