% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectStrategy.R
\name{effectStrategy}
\alias{effectStrategy}
\title{effectStrategy (abstract base class for effect strategies)}
\description{
Defines the interface for effect strategies (PD, ALE, SD, etc.).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-effectStrategy-new}{\code{effectStrategy$new()}}
\item \href{#method-effectStrategy-preprocess}{\code{effectStrategy$preprocess()}}
\item \href{#method-effectStrategy-node_transform}{\code{effectStrategy$node_transform()}}
\item \href{#method-effectStrategy-heterogeneity}{\code{effectStrategy$heterogeneity()}}
\item \href{#method-effectStrategy-find_best_split}{\code{effectStrategy$find_best_split()}}
\item \href{#method-effectStrategy-clone}{\code{effectStrategy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-effectStrategy-new"></a>}}
\if{latex}{\out{\hypertarget{method-effectStrategy-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{effectStrategy$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-effectStrategy-preprocess"></a>}}
\if{latex}{\out{\hypertarget{method-effectStrategy-preprocess}{}}}
\subsection{Method \code{preprocess()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{effectStrategy$preprocess(
  effect,
  data,
  target.feature,
  feature.set = NULL,
  split.feature = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-effectStrategy-node_transform"></a>}}
\if{latex}{\out{\hypertarget{method-effectStrategy-node_transform}{}}}
\subsection{Method \code{node_transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{effectStrategy$node_transform(Y, grid, idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-effectStrategy-heterogeneity"></a>}}
\if{latex}{\out{\hypertarget{method-effectStrategy-heterogeneity}{}}}
\subsection{Method \code{heterogeneity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{effectStrategy$heterogeneity(Y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-effectStrategy-find_best_split"></a>}}
\if{latex}{\out{\hypertarget{method-effectStrategy-find_best_split}{}}}
\subsection{Method \code{find_best_split()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{effectStrategy$find_best_split(Z, Y, min.node.size, n.quantiles)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-effectStrategy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-effectStrategy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{effectStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
