% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_general.R
\name{extract_split_criteria}
\alias{extract_split_criteria}
\title{Print summary of split criteria in a partition tree}
\usage{
extract_split_criteria(tree, feat_name = NULL)
}
\arguments{
\item{tree}{A list of tree nodes structured by depth,}

\item{feat_name}{Character string specifying a single feature name to filter the
displayed heterogeneity (`objective.value`).}
}
\value{
No return value. The function prints to console.
}
\description{
Print summary of split criteria in a partition tree
}
