% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_to_numeric.R
\name{factor_to_numeric}
\alias{factor_to_numeric}
\title{Convert factor to numeric (level labels or level indices)}
\usage{
factor_to_numeric(f)
}
\arguments{
\item{f}{Factor.}
}
\value{
Numeric vector.
}
\description{
If all level labels are numeric strings, converts via \code{as.character};
otherwise uses \code{as.numeric(f)} (level indices).
}
\keyword{internal}
