% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_parent_by_id.R
\name{find_parent_by_id}
\alias{find_parent_by_id}
\title{Find a node by id in a flat list of nodes}
\usage{
find_parent_by_id(node.list, id)
}
\arguments{
\item{node.list}{List of node objects (each with an \code{id} field).}

\item{id}{Scalar; node id to look up.}
}
\value{
The node with \code{id} or \code{NULL}.
}
\description{
Find a node by id in a flat list of nodes
}
\keyword{internal}
