% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gp.R
\name{fit_gp}
\alias{fit_gp}
\title{Fit a Gaussian process to the design criterion surface}
\usage{
fit_gp(experiment, design, response)
}
\arguments{
\item{experiment}{GADGET experiment object}

\item{design}{a matrix of design points}

\item{response}{a column of responses}
}
\description{
The next batch of design points in the experiment are chosen by optimizing the
design criterion over the batch design space. This function updates the Gaussian
process model which is used by EQI to sequentially find the best batch design
given the current design and prior distribution.
}

