% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GADGET.R
\docType{package}
\name{GADGET}
\alias{GADGET}
\alias{GADGET-package}
\title{Gaussian Process Approximations for Designing Experiment}
\description{
The \code{GADGET} package offers
}
\references{
Weaver, B. P., Williams, B. J., Anderson-Cook, C. M., Higdon, D. M. (2016). Computational enhancements to Bayesian design of experiments using Gaussian processes. Bayesian Analysis, 11(1), 191–213, <doi:10.1214/15-BA945>.
}
\author{
Isaac Michaud, Brian Weaver, and Brian Williams
}
