% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gadget}
\alias{gadget}
\title{A low level gadget run}
\usage{
gadget(args)
}
\arguments{
\item{args}{A vector of string arguments}
}
\description{
This function initializes gadget using a given string vector argument.
This is equivalent to running gadget from the command line. Please change
the working directory (e.g., by \code{\link{setwd}})to the model directory
before calling this function. Must be followed by \code{\link{initSim}}
}
\examples{
\dontrun{
 if(interactive()){
  exPath <- loadExample()
  setwd(exPath)
  gadget(c("-s", "-main", "main", "-i", "refinputfile"))
  initSim()
 }
}

}
\seealso{
\code{\link{initGadget}} for the recommended way to initialize
gadget.
}
