% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetTree.R
\name{gadgetTree}
\alias{gadgetTree}
\title{gadgetTree: Main tree class for gadget}
\description{
This class implements the main tree growing logic, using a strategy object (e.g., pdStrategy) for effect-specific operations.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gadgetTree-new}{\code{gadgetTree$new()}}
\item \href{#method-gadgetTree-fit}{\code{gadgetTree$fit()}}
\item \href{#method-gadgetTree-plot}{\code{gadgetTree$plot()}}
\item \href{#method-gadgetTree-plot_tree_structure}{\code{gadgetTree$plot_tree_structure()}}
\item \href{#method-gadgetTree-extract_split_info}{\code{gadgetTree$extract_split_info()}}
\item \href{#method-gadgetTree-clone}{\code{gadgetTree$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gadgetTree-new"></a>}}
\if{latex}{\out{\hypertarget{method-gadgetTree-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gadgetTree$new(
  strategy,
  n.split = 2,
  impr.par = 0.1,
  min.node.size = 10,
  n.quantiles = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gadgetTree-fit"></a>}}
\if{latex}{\out{\hypertarget{method-gadgetTree-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gadgetTree$fit(
  effect,
  data,
  target.feature.name,
  feature.set = NULL,
  split.feature = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gadgetTree-plot"></a>}}
\if{latex}{\out{\hypertarget{method-gadgetTree-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gadgetTree$plot(
  effect,
  data,
  target.feature.name,
  depth = NULL,
  node.id = NULL,
  features = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gadgetTree-plot_tree_structure"></a>}}
\if{latex}{\out{\hypertarget{method-gadgetTree-plot_tree_structure}{}}}
\subsection{Method \code{plot_tree_structure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gadgetTree$plot_tree_structure()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gadgetTree-extract_split_info"></a>}}
\if{latex}{\out{\hypertarget{method-gadgetTree-extract_split_info}{}}}
\subsection{Method \code{extract_split_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gadgetTree$extract_split_info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gadgetTree-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gadgetTree-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gadgetTree$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
