% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{initGadget}
\alias{initGadget}
\title{Initialize a gadget simulation}
\usage{
initGadget(modelDir, inputfile, mainfile = "main")
}
\arguments{
\item{modelDir}{Path to the gadget model}

\item{inputfile}{The name of the model parameter input file}

\item{mainfile}{The name of the model main file. Default: "main"}
}
\description{
This function initialize a gadget simulation given a valid model directory and a input file.
The simulation will not be run yet, please see the \code{\link{runStep}} and \code{\link{runYear}} functions.
}
