% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plots.R
\name{mean_center_ice}
\alias{mean_center_ice}
\title{Mean-center ICE curves}
\usage{
mean_center_ice(ice, feature)
}
\arguments{
\item{ice}{A **data.frame** (or **data.table**) in long ICE format;
must contain a value column \code{.value}.}

\item{feature}{Column name (string) used as the key when spreading to
wide format.}
}
\value{
A **data.table** (in wide format) of centered ICE curves
}
\description{
Converts a long‐format ICE data set into a wide matrix
}
\keyword{internal}
