% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdStrategy.R
\name{pdStrategy}
\alias{pdStrategy}
\title{pdStrategy: Partial Dependence Tree Strategy (R6 class)}
\description{
Implements the effectStrategy interface for building and analyzing Partial Dependence (PD) trees.
This strategy supports data preprocessing, node transformation, heterogeneity calculation, best split search, tree fitting, and visualization.
}
\details{
This class is used internally by the gadgetTree framework to implement partial dependence tree growing, splitting, and visualization. It is not intended to be used directly by end users, but can be instantiated for advanced customization.
}
\section{Fields}{

\describe{
  \item{tree_ref}{Reference to the associated tree instance.}
}
}

\section{Methods}{

\describe{
  \item{initialize()}{Initialize the strategy with name "pd".}
  \item{preprocess(effect, data, target.feature, feature.set = NULL, split.feature = NULL)}{
    Preprocess input data to generate the split feature set Z, effect list Y, and grid.
    @param effect (`R6` object or list)\cr
      An object containing feature effect results, typically from `FeatureEffect` or `FeatureEffects`.
    @param data (`data.frame`)\cr
      Data frame containing all features and the target variable.
    @param target.feature (`character(1)`)\cr
      The name of the target feature in the data to explain.
    @param feature.set (`character` or `NULL`)\cr
      Optional. Subset of features to use for effect calculation. If `NULL`, all features are used.
    @param split.feature (`character` or `NULL`)\cr
      Optional. Features to consider for splitting at each node. If `NULL`, all features are considered.
    @return (`list`)\cr
      A list with Z (split feature set), Y (effect list), and grid (feature grid).
  }
  \item{node_transform(Y, grid, idx)}{
    Mean-center the effect list Y within the node and return the centered list.
    @param Y (`list`)\cr
      Each element is an effect matrix for a feature.
    @param grid (`list`)\cr
      Feature grids.
    @param idx (`integer`)\cr
      Sample indices for the current node.
    @return (`list`)\cr
      List of centered effects.
  }
  \item{heterogeneity(Y)}{
    Calculate the heterogeneity (sum of variances) of effects within the node.
    @param Y (`list`)\cr
      List of effect matrices.
    @return (`numeric`)\cr
      Numeric vector, heterogeneity for each feature.
  }
  \item{find_best_split(Z, Y, min.node.size, n.quantiles)}{
    Find the best split point for a node.
    @param Z (`data.frame`)\cr
      Split feature set.
    @param Y (`list`)\cr
      Effect matrices.
    @param min.node.size (`integer(1)`)\cr
      Minimum node size.
    @param n.quantiles (`integer(1)` or `NULL`)\cr
      Optional. Number of quantiles to use for candidate split points (for numeric features). If `NULL`, use default.
    @return (`list`)\cr
      List with best split feature, split point, etc.
  }
  \item{plot(tree, effect, data, target.feature.name, depth = NULL, node.id = NULL, features = NULL, ...)}{
    Visualize the PD tree structure and partial dependence plots for each node.
    @param tree (`list`)\cr
      Tree structure as a list of Node objects.
    @param effect (`R6` object or list)\cr
      Model effect object.
    @param data (`data.frame`)\cr
      Data frame.
    @param target.feature.name (`character(1)`)\cr
      Target feature name.
    @param depth (`integer` or `NULL`)\cr
      Depth(s) to visualize (optional).
    @param node.id (`integer` or `NULL`)\cr
      Node id(s) to visualize (optional).
    @param features (`character` or `NULL`)\cr
      Features to visualize (optional).
    @param ... Additional plotting arguments.
    @return (`list`)\cr
      List of ggplot2 objects for different depths and nodes.
  }
  \item{fit(tree_instance, effect, data, target.feature.name, feature.set = NULL, split.feature = NULL)}{
    Fit a PD tree using the provided data and effect object.
    @param tree_instance (`gadgetTree` object)\cr
      Tree object instance.
    @param effect (`R6` object or list)\cr
      Model effect object.
    @param data (`data.frame`)\cr
      Data frame.
    @param target.feature.name (`character(1)`)\cr
      Target feature name.
    @param feature.set (`character` or `NULL`)\cr
      Feature subset (optional).
    @param split.feature (`character` or `NULL`)\cr
      Split feature (optional).
    @return (`gadgetTree` object, invisibly)\cr
      The fitted tree object.
  }
}
}

\examples{
# Example: Fit and plot a PD tree using pdStrategy and gadgetTree
# (Assuming effect and data are prepared)
pd_strat <- pdStrategy$new()
tree <- gadgetTree$new(strategy = pd_strat, n.split = 2)
tree$fit(effect, data, target.feature.name = "target")
tree$plot(effect, data, target.feature.name = "target")

}
\section{Super class}{
\code{\link[gadget:effectStrategy]{gadget::effectStrategy}} -> \code{pdStrategy}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-pdStrategy-new}{\code{pdStrategy$new()}}
\item \href{#method-pdStrategy-preprocess}{\code{pdStrategy$preprocess()}}
\item \href{#method-pdStrategy-node_transform}{\code{pdStrategy$node_transform()}}
\item \href{#method-pdStrategy-heterogeneity}{\code{pdStrategy$heterogeneity()}}
\item \href{#method-pdStrategy-find_best_split}{\code{pdStrategy$find_best_split()}}
\item \href{#method-pdStrategy-plot}{\code{pdStrategy$plot()}}
\item \href{#method-pdStrategy-fit}{\code{pdStrategy$fit()}}
\item \href{#method-pdStrategy-clone}{\code{pdStrategy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pdStrategy-new"></a>}}
\if{latex}{\out{\hypertarget{method-pdStrategy-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pdStrategy$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pdStrategy-preprocess"></a>}}
\if{latex}{\out{\hypertarget{method-pdStrategy-preprocess}{}}}
\subsection{Method \code{preprocess()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pdStrategy$preprocess(
  effect,
  data,
  target.feature,
  feature.set = NULL,
  split.feature = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pdStrategy-node_transform"></a>}}
\if{latex}{\out{\hypertarget{method-pdStrategy-node_transform}{}}}
\subsection{Method \code{node_transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pdStrategy$node_transform(Y, grid, idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pdStrategy-heterogeneity"></a>}}
\if{latex}{\out{\hypertarget{method-pdStrategy-heterogeneity}{}}}
\subsection{Method \code{heterogeneity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pdStrategy$heterogeneity(Y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pdStrategy-find_best_split"></a>}}
\if{latex}{\out{\hypertarget{method-pdStrategy-find_best_split}{}}}
\subsection{Method \code{find_best_split()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pdStrategy$find_best_split(Z, Y, min.node.size, n.quantiles)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pdStrategy-plot"></a>}}
\if{latex}{\out{\hypertarget{method-pdStrategy-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pdStrategy$plot(
  tree,
  effect,
  data,
  target.feature.name,
  depth = NULL,
  node.id = NULL,
  features = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pdStrategy-fit"></a>}}
\if{latex}{\out{\hypertarget{method-pdStrategy-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pdStrategy$fit(
  tree_instance,
  effect,
  data,
  target.feature.name,
  feature.set = NULL,
  split.feature = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pdStrategy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-pdStrategy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pdStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
