% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdStrategy.R
\name{pdStrategy}
\alias{pdStrategy}
\title{pdStrategy (partial dependence strategy)}
\description{
Implements the effectStrategy interface for partial dependence (PD).
}
\section{Super class}{
\code{\link[gadget:effectStrategy]{gadget::effectStrategy}} -> \code{pdStrategy}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-pdStrategy-new}{\code{pdStrategy$new()}}
\item \href{#method-pdStrategy-preprocess}{\code{pdStrategy$preprocess()}}
\item \href{#method-pdStrategy-node_transform}{\code{pdStrategy$node_transform()}}
\item \href{#method-pdStrategy-heterogeneity}{\code{pdStrategy$heterogeneity()}}
\item \href{#method-pdStrategy-find_best_split}{\code{pdStrategy$find_best_split()}}
\item \href{#method-pdStrategy-plot}{\code{pdStrategy$plot()}}
\item \href{#method-pdStrategy-fit}{\code{pdStrategy$fit()}}
\item \href{#method-pdStrategy-clone}{\code{pdStrategy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pdStrategy-new"></a>}}
\if{latex}{\out{\hypertarget{method-pdStrategy-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pdStrategy$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pdStrategy-preprocess"></a>}}
\if{latex}{\out{\hypertarget{method-pdStrategy-preprocess}{}}}
\subsection{Method \code{preprocess()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pdStrategy$preprocess(
  effect,
  data,
  target.feature,
  feature.set = NULL,
  split.feature = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pdStrategy-node_transform"></a>}}
\if{latex}{\out{\hypertarget{method-pdStrategy-node_transform}{}}}
\subsection{Method \code{node_transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pdStrategy$node_transform(Y, grid, idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pdStrategy-heterogeneity"></a>}}
\if{latex}{\out{\hypertarget{method-pdStrategy-heterogeneity}{}}}
\subsection{Method \code{heterogeneity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pdStrategy$heterogeneity(Y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pdStrategy-find_best_split"></a>}}
\if{latex}{\out{\hypertarget{method-pdStrategy-find_best_split}{}}}
\subsection{Method \code{find_best_split()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pdStrategy$find_best_split(Z, Y, min.node.size, n.quantiles)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pdStrategy-plot"></a>}}
\if{latex}{\out{\hypertarget{method-pdStrategy-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pdStrategy$plot(
  tree,
  effect,
  data,
  target.feature.name,
  depth = NULL,
  node.id = NULL,
  features = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pdStrategy-fit"></a>}}
\if{latex}{\out{\hypertarget{method-pdStrategy-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pdStrategy$fit(
  tree_instance,
  effect,
  data,
  target.feature.name,
  feature.set = NULL,
  split.feature = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pdStrategy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-pdStrategy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pdStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
