% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regional_pd.R
\name{plot_regional_pd}
\alias{plot_regional_pd}
\title{Plot regional PD/ICE for one node (internal).}
\usage{
plot_regional_pd(
  prepared.data,
  origin.data,
  target.feature.name,
  node.idx,
  color.ice,
  color.pd,
  ymin,
  ymax,
  split.condition = NULL,
  show.point,
  mean.center
)
}
\arguments{
\item{prepared.data, origin.data, target.feature.name, node.idx, color.ice, color.pd, ymin, ymax, split.condition, show.point, mean.center}{Plot arguments.}
}
\description{
Plot regional PD/ICE for one node (internal).
}
\keyword{internal}
