% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_pd.R
\name{plot_tree_pd}
\alias{plot_tree_pd}
\title{Plot PD tree by depth and node (internal).}
\usage{
plot_tree_pd(
  tree,
  effect,
  data,
  target.feature.name,
  color.ice = "lightblue",
  color.pd = "lightcoral",
  show.plot = TRUE,
  show.point = FALSE,
  mean.center = TRUE,
  depth = NULL,
  node.id = NULL,
  features = NULL
)
}
\arguments{
\item{tree, effect, data, target.feature.name, color.ice, color.pd, show.plot, show.point, mean.center, depth, node.id, features}{Plot arguments.}
}
\description{
Plot PD tree by depth and node (internal).
}
\keyword{internal}
