% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_structure.R
\name{plot_tree_structure}
\alias{plot_tree_structure}
\title{Visualize the Tree Structure}
\usage{
plot_tree_structure(tree)
}
\arguments{
\item{tree}{List. A list containing the tree structure, typically converted from the internal Node object format.}
}
\value{
A `ggplot` object representing the tree structure.
}
\description{
This function visualizes the decision tree structure using the `ggraph` and `igraph` packages.
It creates a plot where nodes are labeled with split information and edges represent the tree hierarchy.
}
\keyword{internal}
