% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_plot_data_pd.R
\name{prepare_plot_data_pd}
\alias{prepare_plot_data_pd}
\title{Prepare PD plot data per node (internal).}
\usage{
prepare_plot_data_pd(effect, tree, depth, feature.set, mean.center)
}
\arguments{
\item{effect, tree, depth, feature.set, mean.center}{See PD plotting helpers.}
}
\description{
Prepare PD plot data per node (internal).
}
\keyword{internal}
