% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_pd.R
\name{preprocess_node_data}
\alias{preprocess_node_data}
\title{Preprocess PD node data by depth (internal).}
\usage{
preprocess_node_data(tree, Y, grid.total, mean.center)
}
\description{
Preprocess PD node data by depth (internal).
}
\keyword{internal}
