% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_mean_center_ice.R
\name{re_mean_center_ice}
\alias{re_mean_center_ice}
\title{mean-centering of ICE matrices}
\usage{
re_mean_center_ice(Y, grid, idx)
}
\arguments{
\item{Y}{List of data.frame (one per feature)}

\item{grid}{List of grid column names, i.e. character vectors (one per feature)}

\item{idx}{Integer vector of row indices to keep}
}
\value{
List of mean-centered matrices
}
\description{
mean-centering of ICE matrices
}
