% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_best_split_ale.R
\name{search_best_split_ale}
\alias{search_best_split_ale}
\title{Find best ALE split (internal).}
\usage{
search_best_split_ale(
  Z,
  effect,
  min.node.size = 1L,
  n.quantiles = NULL,
  with_stab = FALSE
)
}
\arguments{
\item{Z, effect, min.node.size, n.quantiles, with_stab}{Split/search arguments.}
}
\description{
Find best ALE split (internal).
}
\keyword{internal}
