% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_best_split_point_ale_with.R
\name{search_best_split_point_ale_with}
\alias{search_best_split_point_ale_with}
\title{Find best ALE split point with stabilizer (internal).}
\usage{
search_best_split_point_ale_with(
  z,
  effect,
  st.table,
  split.feat,
  is.categorical,
  n.quantiles = NULL,
  min.node.size = 1L
)
}
\arguments{
\item{z, effect, st.table, split.feat, is.categorical, n.quantiles, min.node.size}{Arguments.}
}
\description{
Find best ALE split point with stabilizer (internal).
}
\keyword{internal}
