% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_fill.R
\name{space_fill}
\alias{space_fill}
\title{Create space-filling design}
\usage{
space_fill(experiment)
}
\arguments{
\item{experiment}{a GADGET experiment object}
}
\description{
Each stage of a GADGET experiment starts by estimate the current design criterion
surface using a latin hypercude sample (LHS). This function computes the multidimensional
LHS given the initial design budget and the boundary constraints of the design space.
}
