% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ale_heterogeneity.R
\name{ss_ale_dt}
\alias{ss_ale_dt}
\title{Sum-of-squares ALE heterogeneity from list of effect data.tables (internal).}
\usage{
ss_ale_dt(y_list)
}
\arguments{
\item{y_list}{List of ALE data.tables per feature.}
}
\description{
Sum-of-squares ALE heterogeneity from list of effect data.tables (internal).
}
\keyword{internal}
